+++ core/golang_internal_testsuite/testsuite_impl/bulk_command_execution_test/bulk_command_execution_test_.go
@@ -69,12 +69,12 @@ func generateBulkCommandJson(datastoreServiceImage string) string {
                 }
             },
             {
-                "type": "WAIT_FOR_ENDPOINT_AVAILABILITY",
+                "type": "WAIT_FOR_HTTP_GET_ENDPOINT_AVAILABILITY",
                 "args": {
                     "service_id": "service1",
                     "port": 1323,
                     "path": "health",
-                    "initial_delay_seconds": 0,
+                    "initial_delay_milliseconds": 0,
                     "retries": 5,
                     "retries_delay_milliseconds": 2000,
                     "body_text": "healthy"
@@ -98,12 +98,12 @@ func generateBulkCommandJson(datastoreServiceImage string) string {
                 }
             },
             {
-                "type": "WAIT_FOR_ENDPOINT_AVAILABILITY",
+                "type": "WAIT_FOR_HTTP_GET_ENDPOINT_AVAILABILITY",
                 "args": {
                     "service_id": "service2",
                     "port": 1323,
                     "path": "health",
-                    "initial_delay_seconds": 0,
+                    "initial_delay_milliseconds": 0,
                     "retries": 5,
                     "retries_delay_milliseconds": 2000,
                     "body_text": "healthy"
@@ -148,12 +148,12 @@ func generateBulkCommandJson(datastoreServiceImage string) string {
                 }
             },
             {
-                "type": "WAIT_FOR_ENDPOINT_AVAILABILITY",
+                "type": "WAIT_FOR_HTTP_GET_ENDPOINT_AVAILABILITY",
                 "args": {
                     "service_id": "service3",
                     "port": 1323,
                     "path": "health",
-                    "initial_delay_seconds": 0,
+                    "initial_delay_milliseconds": 0,
                     "retries": 5,
                     "retries_delay_milliseconds": 2000,
                     "body_text": "healthy"