+++ core/golang_internal_testsuite/testsuite_impl/local_static_file_test/local_static_file_test_.go
@@ -9,7 +9,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis-client/golang/lib/networks"
 	"github.com/kurtosis-tech/kurtosis-client/golang/lib/services"
 	"github.com/kurtosis-tech/kurtosis-testsuite-api-lib/golang/lib/testsuite"
-	"github.com/kurtosis-tech/kurtosis/internal_testsuite/static_files_consts"
+	"github.com/kurtosis-tech/kurtosis/golang_internal_testsuite/static_files_consts"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 )